requirements.txt
fastapi
uvicorn
python-dotenv
gtts
requests
google-generativeai
jinja2
python-multipart
google-cloud-speech
numpy
pymupdf
bangla
python-docx
